[ENHANCEMENT]: GUN CLASSES AND UPGRADES ARE COMPLETE, PROJECTILE CLASS ORGANIZATION, SPAWNER LOGIC ORGANIZATION.

+++ src/main/java/mapobjects/component/Direction.java
@@ -4,7 +4,7 @@ public class Direction {
 
     private double direction;
 
-    public Direction(int direction) {
+    public Direction(double direction) {
         this.direction = Math.toRadians(direction);
     }
